@pep8speaks 2018-01-04
Hello @toobaz! Thanks for submitting the PR.

 - In the file [`pandas/core/indexes/multi.py`](https://github.com/pandas-dev/pandas/blob/6e1ecec1c4d9e1926eeb3caad1203a1cc3f4fb74/pandas/core/indexes/multi.py), following are the PEP8 issues :

> [Line 117:9](https://github.com/pandas-dev/pandas/blob/6e1ecec1c4d9e1926eeb3caad1203a1cc3f4fb74/pandas/core/indexes/multi.py#L117): [E722](https://duckduckgo.com/?q=pep8%20E722) do not use bare except'