@jreback 2018-01-04
> .. clearly show that the engine increases performance for large indexes, and significantly decreases performance for get_loc on small/medium indexes. For reference, on the small test case (10 elements), 16 μs. are lost on each get_loc call, while on the large test case (1 million elements), 146 μs. are gained on each get_loc call. I think the tradeoff is acceptable, also considering that 

This is exactly why we changed this previously. Please look at the previous issue and how @jorisvandenbossche made the tradeoff chart. 

This needs a full asv run as well as significant code cleanup. But first if you can narrow the small indexing perf then that would be helpful.