@toobaz 2018-01-04
> This is exactly why we changed this previously. Please look at the previous issue

What do you mean by "this"? I don't think pandas ever had a ``MultiIndex`` engine similar to the one I am proposing. I saw #16319, but there we were talking about [a ~x1000](https://github.com/pandas-dev/pandas/issues/16319#issuecomment-300533260), not ~x5, performance impact.

> and how @jorisvandenbossche made the tradeoff chart.

pointer?

> But first if you can narrow the small indexing perf then that would be helpful.

What do you mean?