@TomAugspurger 2018-01-04
> pointer?


https://github.com/pandas-dev/pandas/pull/16324#issuecomment-300806285