@toobaz 2018-01-04
> #16324 (comment)

Thanks, had missed that! So here is a comparison of the three engines:
http://nbviewer.jupyter.org/gist/toobaz/de76b3fbc6745266fd4c2a6e428bd2a4

And here is the figure equivalent to that in https://github.com/pandas-dev/pandas/pull/16324#issuecomment-300806285 :

![perf_warm_w_10](https://user-images.githubusercontent.com/1224492/34572291-504bb40e-f171-11e7-9253-f2c4e372471e.png)