@jreback 2018-01-04
ok from a real world perspective then for < 1000 is slower, but how much in total time?