@toobaz 2018-01-04
> ok from a real world perspective then for < 1000 is slower, but how much in total time?

Assuming you are asking for the difference between UInt and Object in the "warm combination" case (dashed lines), we are in the order of 100 μs:

```
9        0.000178
100      0.000147
900      0.000059
10000   -0.000757
```

(this is tested with more iterations than the figure above, see the [updated notebook](https://gist.github.com/toobaz/de76b3fbc6745266fd4c2a6e428bd2a4))