@toobaz 2018-01-04
@jbrockmendel's suggestion to cythonize did help reducing a bit the performance decrease:
```
       before           after         ratio
     [93033151]       [b3ba47c6]
+     3.97±0.02μs      17.4±0.04μs     4.39  multiindex_object.GetLoc.time_med_get_loc
+     4.13±0.02ms      18.0±0.08ms     4.36  multiindex_object.GetLoc.time_small_get_loc_warm
+     4.26±0.03μs       17.6±0.1μs     4.13  multiindex_object.GetLoc.time_string_get_loc
+     4.32±0.02ms      16.6±0.05ms     3.84  multiindex_object.GetLoc.time_med_get_loc_warm
-         181±5ms        153±0.8ms     0.84  multiindex_object.GetLoc.time_large_get_loc
-         170±2ms        123±0.5ms     0.73  multiindex_object.Integer.time_get_indexer
-       321±0.8ms        169±0.6ms     0.53  multiindex_object.GetLoc.time_large_get_loc_warm

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
```