@toobaz 2018-01-11
New asv run shows some small additional improvement:
```
       before           after         ratio
     [9d8dbefb]       [b152cb80]
+     4.71±0.07μs       19.9±0.5μs     4.21  multiindex_object.GetLoc.time_string_get_loc
+     4.69±0.05ms       18.2±0.2ms     3.88  multiindex_object.GetLoc.time_small_get_loc_warm
+      5.80±0.9μs       19.8±0.1μs     3.41  multiindex_object.GetLoc.time_med_get_loc
+      5.86±0.5ms       18.7±0.4ms     3.19  multiindex_object.GetLoc.time_med_get_loc_warm
+      12.8±0.1ms      18.2±0.05ms     1.42  multiindex_object.Values.time_datetime_level_values_copy
-         464±8μs          421±5μs     0.91  multiindex_object.Duplicates.time_remove_unused_levels
-         201±7ms        165±0.9ms     0.82  multiindex_object.GetLoc.time_large_get_loc
-         197±4ms        132±0.4ms     0.67  multiindex_object.Integer.time_get_indexer
-        426±20ms          182±2ms     0.43  multiindex_object.GetLoc.time_large_get_loc_warm

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
```