@jorisvandenbossche 2018-01-12
@toobaz Do you think there is further potential in optimizing get_loc to reduce the performance degradation?

Performance is important, but correct and consistent (for different MI sizes) behaviour as well. And it seems this change improves a lot on that aspect (didn't look in detail at the the actual diff).