@toobaz 2018-01-12
> @toobaz Do you think there is further potential in optimizing get_loc to reduce the performance degradation?
> 
> Performance is important, but correct and consistent (for different MI sizes) behaviour as well. And it seems this change improves a lot on that aspect (didn't look in detail at the the actual diff).

Performance increase for large index is in absolute value way bigger than the performance loss on small ones, so even just performance-wise, I think this PR is an improvement. pandas (and numpy, by the way) users (should) expect good performance for big stuff and overhead for small stuff.

Then clearly if there was a way to reduce overhead, it would be worth doing. But the only new code which does take some cycles (as confirmed by profiling, before cythonizing) is ``codes_to_ints``, where I don't see room of improvement. Real improvement could come from having ``Index.get_loc`` not recur to ``Index.get_indexer``, but 1) I'm not sure 2) it could vary from class to class 3) anyway, could and should be done separately