@jorisvandenbossche 2018-01-12
> Real improvement could come from having Index.get_loc

According to my profiling, the individual `get_loc` calls on the levels only account for around 20% of the time, so it will only give modest improvements for this case (but of course, for indexing in general that would certainly be welcome!)

Anyhow, as I said, I agree this is the other improvements here make it worth it!