@toobaz 2018-01-12
> if something is a helper method that is only used within the class, make this clear by using a _... name

Yeah, and the same holds for ``base`` vs. ``_base``... I think these simple conventions make code more readable than tons of docstrings nobody will read (not talking about the API, obviously).