@jreback 2018-01-17
can you re-run and update the benchmarks at the very top of this PR (after you make changes).