@toobaz 2018-01-17
New changes reduced the overhead further...
```
       before           after         ratio
     [8acdf801]       [bbae39c8]
+      4.41±0.1ms       15.4±0.1ms     3.49  multiindex_object.GetLoc.time_small_get_loc_warm
+     4.73±0.06ms       15.1±0.5ms     3.19  multiindex_object.GetLoc.time_med_get_loc_warm
+     4.45±0.07μs       13.9±0.2μs     3.13  multiindex_object.GetLoc.time_string_get_loc
+     4.75±0.03μs      14.8±0.05μs     3.12  multiindex_object.GetLoc.time_med_get_loc
-         168±1ms          148±1ms     0.89  multiindex_object.Integer.time_get_indexer
-         183±6ms          156±3ms     0.85  multiindex_object.GetLoc.time_large_get_loc
-         349±5ms          169±2ms     0.49  multiindex_object.GetLoc.time_large_get_loc_warm

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
```

... but trying to ``cpdef`` any of the methods of the cython mother class (or ``cdef _extract_level_codes``, which is only used internally) as suggested by @jbrockmendel results in

``` python
Traceback (most recent call last):
  File "/usr/lib/python3/dist-packages/_pytest/config.py", line 379, in _importconftest
    mod = conftestpath.pyimport()
  File "/usr/lib/python3/dist-packages/py/_path/local.py", line 662, in pyimport
    __import__(modname)
  File "/home/nobackup/repo/pandas/pandas/__init__.py", line 42, in <module>
    from pandas.core.api import *
  File "/home/nobackup/repo/pandas/pandas/core/api.py", line 10, in <module>
    from pandas.core.groupby import Grouper
  File "/home/nobackup/repo/pandas/pandas/core/groupby.py", line 48, in <module>
    from pandas.core.index import (Index, MultiIndex,
  File "/home/nobackup/repo/pandas/pandas/core/index.py", line 2, in <module>
    from pandas.core.indexes.api import *
  File "/home/nobackup/repo/pandas/pandas/core/indexes/api.py", line 8, in <module>
    from pandas.core.indexes.multi import MultiIndex  # noqa
  File "/home/nobackup/repo/pandas/pandas/core/indexes/multi.py", line 54, in <module>
    libindex.UInt64Engine):
TypeError: multiple bases have instance lay-out conflict
ERROR: could not load /home/nobackup/repo/pandas/pandas/conftest.py

```